Функциональные эндпоинты. Перевел хэндлеры на интерфейсы, вынес листы в переменные.
